+++ scry/src/price/models.rs
@@ -5,7 +5,7 @@ use sqlx::FromRow;
 
 #[derive(Debug, FromRow, Serialize, Deserialize)]
 pub struct Price {
-    pub id: Option<i64>,
+    pub id: Option<i32>,
     pub card_id: String,
     pub foil: Option<Decimal>,
     pub normal: Option<Decimal>,
